create /constants
