+++ components/PostsBoardNavbar.tsx
@@ -1,13 +1,5 @@
 import Link from "next/link";
-
-export const boardList = [
-  { id: 0, name: "피드", route: "" },
-  { id: 1, name: "익명", route: "anonymous" },
-  { id: 2, name: "자유", route: "free" },
-  { id: 3, name: "구인", route: "career" },
-  { id: 4, name: "정보", route: "information" },
-  { id: 5, name: "홍보", route: "ad" },
-];
+import { boardList } from "../constants/constants";
 
 export const PostsBoardNavbar = () => {
   const buttonList = boardList.map((board) => {
