+++ constants/constants.ts
@@ -1,0 +1,11 @@
+export const boardList = [
+  { id: 0, name: "피드", route: "" },
+  { id: 1, name: "익명", route: "anonymous" },
+  { id: 2, name: "자유", route: "free" },
+  { id: 3, name: "구인", route: "career" },
+  { id: 4, name: "정보", route: "information" },
+  { id: 5, name: "홍보", route: "ad" },
+];
+
+export const LOADING = "loading";
+export const ERROR = "error";
